+++ install-scripts/install-macos.sh
@@ -15,3 +15,10 @@ brew install pyenv gmp
 pyenv -v
 
 make deps-macos
+
+pyenv local 3.9.15
+
+pip install -r requirements.txt
+
+echo "-- You need to follow these instructions to finish installing pyenv: --"
+pyenv init || true