+++ install-scripts/install-ubuntu.sh
@@ -24,8 +24,14 @@ export PYENV_ROOT="$HOME/.pyenv"
 command -v pyenv >/dev/null || export PATH="$PYENV_ROOT/bin:$PATH"
 eval "$(pyenv init -)"
 
-
 # Make sure pyenv has been installed correctly
 pyenv -v
 
 make deps
+
+pyenv local 3.9.15
+
+pip install -r requirements.txt
+
+echo "-- You need to follow these instructions to finish installing pyenv: --"
+pyenv init || true