+++ vm/src/hint_processor/cairo_1_hint_processor/hint_processor.rs
@@ -526,9 +526,9 @@ impl Cairo1HintProcessor {
         let x_bigint: BigUint = random_x.into_bigint().into();
         let y_bigint: BigUint = random_y_squared
             .sqrt()
-            .ok_or(HintError::CustomHint(
-                "Failed to compute sqrt".to_string().into_boxed_str(),
-            ))?
+            .ok_or_else(|| {
+                HintError::CustomHint("Failed to compute sqrt".to_string().into_boxed_str())
+            })?
             .into_bigint()
             .into();
 
@@ -568,11 +568,13 @@ impl Cairo1HintProcessor {
             .get_integer((dict_manager_address - 2)?)?
             .into_owned()
             .to_usize()
-            .ok_or(HintError::CustomHint(
-                "Invalid number of dictionaries."
-                    .to_string()
-                    .into_boxed_str(),
-            ))?;
+            .ok_or_else(|| {
+                HintError::CustomHint(
+                    "Invalid number of dictionaries."
+                        .to_string()
+                        .into_boxed_str(),
+                )
+            })?;
 
         let dict_infos_base = vm.get_relocatable((dict_manager_address - 3)?)?;
 
@@ -832,9 +834,9 @@ impl Cairo1HintProcessor {
         let val = Felt252::from(
             if dict_squash_exec_scope
                 .current_access_indices()
-                .ok_or(HintError::CustomHint(
-                    "no indices accessed".to_string().into_boxed_str(),
-                ))?
+                .ok_or_else(|| {
+                    HintError::CustomHint("no indices accessed".to_string().into_boxed_str())
+                })?
                 .len()
                 > 1
             {
@@ -882,17 +884,14 @@ impl Cairo1HintProcessor {
     ) -> Result<(), HintError> {
         let dict_squash_exec_scope: &mut DictSquashExecScope =
             exec_scopes.get_mut_ref("dict_squash_exec_scope")?;
-        let prev_access_index =
-            dict_squash_exec_scope
-                .pop_current_access_index()
-                .ok_or(HintError::CustomHint(
-                    "no accessed index".to_string().into_boxed_str(),
-                ))?;
+        let prev_access_index = dict_squash_exec_scope
+            .pop_current_access_index()
+            .ok_or_else(|| {
+                HintError::CustomHint("no accessed index".to_string().into_boxed_str())
+            })?;
         let index_delta_minus_1_val = dict_squash_exec_scope
             .current_access_index()
-            .ok_or(HintError::CustomHint(
-                "no index accessed".to_string().into_boxed_str(),
-            ))?
+            .ok_or_else(|| HintError::CustomHint("no index accessed".to_string().into_boxed_str()))?
             .clone()
             - prev_access_index
             - 1_u32;
@@ -925,14 +924,15 @@ impl Cairo1HintProcessor {
             exec_scopes.get_mut_ref::<DictSquashExecScope>("dict_squash_exec_scope")?;
         let (dict_accesses_base, dict_accesses_offset) = extract_buffer(dict_accesses)?;
         let dict_accesses_address = get_ptr(vm, dict_accesses_base, &dict_accesses_offset)?;
-        let n_accesses =
-            res_operand_get_val(vm, n_accesses)?
-                .to_usize()
-                .ok_or(HintError::CustomHint(
+        let n_accesses = res_operand_get_val(vm, n_accesses)?
+            .to_usize()
+            .ok_or_else(|| {
+                HintError::CustomHint(
                     "Number of accesses is too large or negative."
                         .to_string()
                         .into_boxed_str(),
-                ))?;
+                )
+            })?;
 
         for i in 0..n_accesses {
             let current_key = vm.get_integer((dict_accesses_address + i * dict_access_size)?)?;
@@ -966,11 +966,9 @@ impl Cairo1HintProcessor {
 
         vm.insert_value(
             cell_ref_to_relocatable(first_key, vm)?,
-            dict_squash_exec_scope
-                .current_key()
-                .ok_or(HintError::CustomHint(
-                    "No current key".to_string().into_boxed_str(),
-                ))?,
+            dict_squash_exec_scope.current_key().ok_or_else(|| {
+                HintError::CustomHint("No current key".to_string().into_boxed_str())
+            })?,
         )?;
 
         Ok(())
@@ -1022,9 +1020,9 @@ impl Cairo1HintProcessor {
         let current_access_index =
             dict_squash_exec_scope
                 .current_access_index()
-                .ok_or(HintError::CustomHint(
-                    "No current accessed index".to_string().into_boxed_str(),
-                ))?;
+                .ok_or_else(|| {
+                    HintError::CustomHint("No current accessed index".to_string().into_boxed_str())
+                })?;
         vm.insert_value(range_check_ptr, current_access_index)?;
 
         Ok(())
@@ -1115,7 +1113,7 @@ impl HintProcessorLogic for Cairo1HintProcessor {
         _references: &[HintReference],
     ) -> Result<Box<dyn Any>, VirtualMachineError> {
         let data = hint_code.parse().ok().and_then(|x: usize| self.hints.get(&x).cloned())
-        .ok_or(VirtualMachineError::CompileHintFail(
+        .ok_or_else(|| VirtualMachineError::CompileHintFail(
             format!("No hint found for pc {hint_code}. Cairo1HintProccesor can only be used when running CasmContractClass").into_boxed_str()
     ))?;
         Ok(any_box!(data))