+++ vm/src/hint_processor/cairo_1_hint_processor/hint_processor.rs
@@ -27,7 +27,7 @@ use core::ops::Shl;
 
 use num_bigint::BigUint;
 use num_integer::Integer;
-use num_traits::{cast::ToPrimitive, Zero};
+use num_traits::cast::ToPrimitive;
 
 /// Execution scope for constant memory allocation.
 struct MemoryExecScope {
@@ -124,32 +124,18 @@ impl Cairo1HintProcessor {
             }
 
             Hint::Core(CoreHintBase::Core(CoreHint::Uint256DivMod {
-                dividend_low,
-                dividend_high,
-                divisor_low,
-                divisor_high,
-                quotient0,
-                quotient1,
+                dividend0,
+                dividend1,
                 divisor0,
                 divisor1,
-                extra0,
-                extra1,
-                remainder_low,
-                remainder_high,
+                quotient0,
+                quotient1,
+
+                remainder0,
+                remainder1,
             })) => self.uint256_div_mod(
-                vm,
-                dividend_low,
-                dividend_high,
-                divisor_low,
-                divisor_high,
-                quotient0,
-                quotient1,
-                divisor0,
-                divisor1,
-                extra0,
-                extra1,
-                remainder_low,
-                remainder_high,
+                vm, dividend0, dividend1, divisor0, divisor1, quotient0, quotient1, remainder0,
+                remainder1,
             ),
             Hint::Core(CoreHintBase::Deprecated(DeprecatedHint::Felt252DictWrite {
                 dict_ptr,
@@ -410,64 +396,42 @@ impl Cairo1HintProcessor {
     fn uint256_div_mod(
         &self,
         vm: &mut VirtualMachine,
-        dividend_low: &ResOperand,
-        dividend_high: &ResOperand,
-        divisor_low: &ResOperand,
-        divisor_high: &ResOperand,
+        dividend0: &ResOperand,
+        dividend1: &ResOperand,
+        divisor0: &ResOperand,
+        divisor1: &ResOperand,
         quotient0: &CellRef,
         quotient1: &CellRef,
-        divisor0: &CellRef,
-        divisor1: &CellRef,
-        extra0: &CellRef,
-        extra1: &CellRef,
-        remainder_low: &CellRef,
-        remainder_high: &CellRef,
+        remainder0: &CellRef,
+        remainder1: &CellRef,
     ) -> Result<(), HintError> {
         let pow_2_128 = BigUint::from(u128::MAX) + 1u32;
-        let pow_2_64 = BigUint::from(u64::MAX) + 1u32;
-        let dividend_low = res_operand_get_val(vm, dividend_low)?.to_biguint();
-        let dividend_high = res_operand_get_val(vm, dividend_high)?.to_biguint();
-        let divisor_low = res_operand_get_val(vm, divisor_low)?.to_biguint();
-        let divisor_high = res_operand_get_val(vm, divisor_high)?.to_biguint();
-        let dividend = dividend_low + dividend_high * &pow_2_128;
-        let divisor = divisor_low + &divisor_high * &pow_2_128;
-        let quotient = &dividend / &divisor;
-        let remainder = dividend % &divisor;
-
-        // Guess quotient limbs.
-        let (quotient, limb) = quotient.div_rem(&pow_2_64);
-        vm.insert_value(cell_ref_to_relocatable(quotient0, vm)?, Felt252::from(limb))?;
-        let (quotient, limb) = quotient.div_rem(&pow_2_64);
-        vm.insert_value(cell_ref_to_relocatable(quotient1, vm)?, Felt252::from(limb))?;
-        let (quotient, limb) = quotient.div_rem(&pow_2_64);
-        if divisor_high.is_zero() {
-            vm.insert_value(cell_ref_to_relocatable(extra0, vm)?, Felt252::from(limb))?;
-            vm.insert_value(
-                cell_ref_to_relocatable(extra1, vm)?,
-                Felt252::from(quotient),
-            )?;
-        }
-
-        // Guess divisor limbs.
-        let (divisor, limb) = divisor.div_rem(&pow_2_64);
-        vm.insert_value(cell_ref_to_relocatable(divisor0, vm)?, Felt252::from(limb))?;
-        let (divisor, limb) = divisor.div_rem(&pow_2_64);
-        vm.insert_value(cell_ref_to_relocatable(divisor1, vm)?, Felt252::from(limb))?;
-        let (divisor, limb) = divisor.div_rem(&pow_2_64);
-        if !divisor_high.is_zero() {
-            vm.insert_value(cell_ref_to_relocatable(extra0, vm)?, Felt252::from(limb))?;
-            vm.insert_value(cell_ref_to_relocatable(extra1, vm)?, Felt252::from(divisor))?;
-        }
-
-        // Guess remainder limbs.
+        let dividend0 = get_val(vm, dividend0)?.to_biguint();
+        let dividend1 = get_val(vm, dividend1)?.to_biguint();
+        let divisor0 = get_val(vm, divisor0)?.to_biguint();
+        let divisor1 = get_val(vm, divisor1)?.to_biguint();
+        let dividend: BigUint = dividend0 + dividend1.shl(128);
+        let divisor = divisor0 + divisor1.shl(128);
+        let (quotient, remainder) = dividend.div_rem(&divisor);
+        let (limb1, limb0) = quotient.div_rem(&pow_2_128);
         vm.insert_value(
-            cell_ref_to_relocatable(remainder_low, vm)?,
-            Felt252::from(remainder.clone() % pow_2_128.clone()),
+            cell_ref_to_relocatable(quotient0, vm)?,
+            Felt252::from(limb0),
         )?;
         vm.insert_value(
-            cell_ref_to_relocatable(remainder_high, vm)?,
-            Felt252::from(remainder / pow_2_128),
+            cell_ref_to_relocatable(quotient1, vm)?,
+            Felt252::from(limb1),
         )?;
+        let (limb1, limb0) = remainder.div_rem(&pow_2_128);
+        vm.insert_value(
+            cell_ref_to_relocatable(remainder0, vm)?,
+            Felt252::from(limb0),
+        )?;
+        vm.insert_value(
+            cell_ref_to_relocatable(remainder1, vm)?,
+            Felt252::from(limb1),
+        )?;
+
         Ok(())
     }