+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * refactor: remove private function `bigint3_pack` in favor of equivalent `BigInt3::pack86` method [#1329](https://github.com/lambdaclass/cairo-vm/pull/1329)
 
+* feat: updated the old WASM example and moved it to [`examples/wasm-demo`](examples/wasm-demo/) [#1315](https://github.com/lambdaclass/cairo-vm/pull/1315)
+
+
 * feat: add `arbitrary` feature to enable arbitrary derive in `Program` and `CairoRunConfig`
 
 * feat(fuzzing): add `arbitrary` feature to enable arbitrary derive in `Program` and `CairoRunConfig` [#1306](https://github.com/lambdaclass/cairo-vm/pull/1306) [#1330](https://github.com/lambdaclass/cairo-vm/pull/1330)